# app/static/swagger.yml
openapi: 3.0.0
info:
  title: MS Discs API
  description: Microservicio para gestión de colección de discos
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://127.0.0.1:5000/api
    description: Servidor de desarrollo

tags:
  - name: Discs
    description: Operaciones relacionadas con discos

paths:
  /discs:
    get:
      tags:
        - Discs
      summary: Obtener lista de discos
      description: Retorna todos los discos habilitados. Puede filtrar por tipo o favoritos
      parameters:
        - name: type
          in: query
          description: Filtrar por tipo de disco (cds, vinyls, cassettes)
          required: false
          schema:
            type: string
            pattern: ^[A-Za-z0-9]+$
          example: "cds"
        - name: favorite
          in: query
          description: Filtrar solo discos favoritos
          required: false
          schema:
            type: string
          example: "true"
      responses:
        '200':
          description: Lista de discos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Disc'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
    
    post:
      tags:
        - Discs
      summary: Crear un nuevo disco
      description: Agrega un nuevo disco a la colección
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Disc'
      responses:
        '200':
          description: Disco creado exitosamente
        '400':
          description: Error de validación o disco ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessError'
              examples:
                disc_exists:
                  value:
                    codigo: EXDYE01
                    mensaje: Disco ya existe

  /discs/{sku}:
    get:
      tags:
        - Discs
      summary: Obtener un disco por SKU
      description: Retorna la información detallada de un disco específico
      parameters:
        - name: sku
          in: path
          description: SKU único del disco
          required: true
          schema:
            type: string
          example: "12345"
      responses:
        '200':
          description: Disco encontrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Disc'
        '400':
          description: Disco no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessError'
              examples:
                disc_not_found:
                  value:
                    codigo: EXDNE01
                    mensaje: Disco no encontrado

components:
  schemas:
    Disc:
      type: object
      properties:
        sku:
          type: string
          description: SKU único del disco
          example: "093624933595"
        name:
          type: string
          description: Nombre del disco
          example: "The Black Parade"
        description:
          type: string
          description: Descripción del disco
          example: "The Black Parade. Publicado en 2022"
        author:
          type: string
          description: Autor o artista del disco
          example: "My Chemical Romance"
        publisher:
          type: string
          description: Compañía discográfica
          example: ""
        yearCreated:
          type: integer
          description: Año de lanzamiento
          example: 2022
        country:
          type: string
          description: País de origen
          example: ""
        images:
          type: array
          description: Lista de URLs de imágenes del disco
          items:
            type: string
            example: "1.jpg"
        categories:
          type: array
          description: Lista de categorías o géneros del disco
          items:
            type: string
            example: "Rock"
        trackList:
          type: array
          description: Lista de canciones del disco
          items:
            type: string
            example: "Welcome to the Black Parade"
        type:
          type: string
          description: Tipo o género del disco
          example: "VINYLS"
        favorite:
          type: boolean
          description: Indica si es un disco favorito
          example: true
        enabled:
          type: boolean
          description: Indica si el disco está habilitado
          example: true
      required:
        - sku
        - name
        - description
        - author
        - publisher
        - yearCreated
        - country
        - images
        - categories
        - trackList
        - type
        - favorite
        - enabled
    
    BusinessError:
      type: object
      properties:
        codigo:
          type: string
          description: Código de error de negocio
          example: "EXDNE01"
        mensaje:
          type: string
          description: Mensaje descriptivo del error
          example: "Disco no encontrado"
      required:
        - codigo
        - mensaje
    
    ValidationError:
      type: object
      properties:
        type:
          type: string
          description: Mensaje de error de validación
          example: "Solo admite letras y numeros"